+++ src/main/kotlin/edu/unh/cs/ai/realtimesearch/environment/racetrack/RaceTrackState.kt
@@ -11,12 +11,12 @@ import edu.unh.cs.ai.realtimesearch.environment.location.Location
  *
  * The actual size and shape of the world is state-independent, so not implemented here
  */
-data class RaceTrackState(val x : Double, val y : Double, val x_speed: Int, val y_speed: Int) : State<RaceTrackState> {
+data class RaceTrackState(val x : Int, val y : Int, val x_speed: Int, val y_speed: Int) : State<RaceTrackState> {
 
     override fun equals(other: Any?): Boolean {
         if(other !is RaceTrackState)
             return false
-        if(other.x.toInt() == x.toInt() && other.y.toInt() == y.toInt()) {
+        if(other.x == x && other.y == y) {
             return true;
         }
         return false;
